Lots of new pages for the request page.
In this commit, you will find all of the small little bits of the form, which are the fields which user can add to their form to further specify which fields they wish to consider.
Most follow a very similar format. The most necessary part is the name of the file, and the name of the input element.